project/requirements.txt
asgiref==3.5.2
autopep8==1.6.0
distlib==0.3.4
Django==4.0.5
django-crispy-forms==1.14.0
djangorestframework==3.13.1
filelock==3.7.1
mysqlclient==2.1.1
Pillow==9.2.0
platformdirs==2.5.2
pycodestyle==2.8.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.1
virtualenv==20.14.1
